# docs/swagger.yaml
definitions:
...
  dto.PageableSkillDto:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.SkillDto'
        type: array
      has_next:
        type: boolean
    type: object
  dto.PageableUserDto:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserSkillDto'
        type: array
      has_next:
        type: boolean
    type: object
...
  dto.RegisterSkillDto:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.RegisterUserDtoIn:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  dto.SkillDto:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.UpdateSkillDto:
    properties:
      name:
        type: string
    type: object
  dto.UserDtoOut:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
  dto.UserSkillDto:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      skills:
        items:
          $ref: '#/definitions/dto.SkillDto'
        type: array
    type: object
info:
  contact: {}
  title: AjudaDev Backend API
  version: "1.0"
paths:
...
  /v1/skill:
    get:
      consumes:
      - application/json
      description: 'Retorna as skills com paginação e busca por prefixo do nome (case
        insensitive: o termo é normalizado para caixa alta antes do LIKE)'
      parameters:
      - description: Prefixo do nome da skill
        in: query
        name: name
        type: string
      - description: Página
        in: query
        name: page
        type: integer
      - description: Limite
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PageableSkillDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lista skills do catálogo
      tags:
      - skills
  /v1/skill/{id}:
    delete:
      description: Arquiva a skill marcando deleted_at e remove as associações skill_users
        dela (a skill some do perfil de todos os usuários)
      parameters:
      - description: ID da skill
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove skill (soft delete)
      tags:
      - skills
    get:
      consumes:
      - application/json
      description: Retorna o detalhe da skill
      parameters:
      - description: ID da skill
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SkillDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Busca skill por id
      tags:
      - skills
    put:
      consumes:
      - application/json
      description: Atualiza o nome da skill (normalizado para caixa alta). Nome já
        usado por outra skill ativa gera conflito.
      parameters:
      - description: ID da skill
        in: path
        name: id
        required: true
        type: string
      - description: Novo nome da skill
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSkillDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegisterSkillDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Renomeia uma skill
      tags:
      - skills
...
  /v1/skill/register:
    post:
      consumes:
      - application/json
      description: Cria uma skill no catálogo. O nome é salvo em caixa alta (normalizado)
        e é único entre skills ativas.
      parameters:
      - description: Dados da skill
        in: body
        name: skill
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterSkillDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterSkillDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Registra uma nova skill
      tags:
      - skills
  /v1/user:
    get:
      consumes:
      - application/json
      description: Retorna os usuários ativos que possuem a skill informada (match
        exato do nome, normalizado para caixa alta) com as skills do perfil de cada
        um. O parâmetro skill é obrigatório.
      parameters:
      - description: Nome da skill (normalizado para caixa alta)
        in: query
        name: skill
        required: true
        type: string
      - description: Página
        in: query
        name: page
        type: integer
      - description: Limite
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PageableUserDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Busca usuários por skill
      tags:
      - users
...
  /v1/user/register:
    post:
      consumes:
      - application/json
      description: Cria um novo usuário no sistema
      parameters:
      - description: Dados do usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserDtoIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserDtoOut'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Registra um novo usuário
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"